FAC-391: A condition to check if the picture is present for deletion was added.

+++ flickit-assessment-users/src/main/java/org/flickit/assessment/users/application/service/expertgroup/UpdateExpertGroupPictureService.java
@@ -33,7 +33,8 @@ public class UpdateExpertGroupPictureService implements UpdateExpertGroupPicture
     public Result update(Param param) {
         validateCurrentUser(param.getExpertGroupId(), param.getCurrentUserId());
         var picture = loadExpertGroupPort.loadExpertGroup(param.getExpertGroupId()).getPicture();
-        deleteFilePort.delete(picture);
+        if (picture != null && !picture.isBlank())
+            deleteFilePort.delete(picture);
 
         picture = uploadExpertGroupPicturePort.uploadPicture(param.getPicture());
         updateExpertGroupPicturePort.updatePicture(param.getExpertGroupId(), picture);
